+++ test/test.js
@@ -458,6 +458,56 @@ describe('finalhandler(req, res)', function () {
     })
   })
 
+  describe('headers set', function () {
+    it('should persist set headers', function (done) {
+      var server = http.createServer(function (req, res) {
+        var done = finalhandler(req, res)
+        res.setHeader('Server', 'foobar')
+        done()
+      })
+
+      request(server)
+        .get('/foo')
+        .expect(404)
+        .expect('Server', 'foobar')
+        .end(done)
+    })
+
+    it('should override content-type and length', function (done) {
+      var server = http.createServer(function (req, res) {
+        var done = finalhandler(req, res)
+        res.setHeader('Content-Type', 'image/png')
+        res.setHeader('Content-Length', '50')
+        done()
+      })
+
+      request(server)
+        .get('/foo')
+        .expect(404)
+        .expect('Content-Type', 'text/html; charset=utf-8')
+        .expect('Content-Length', '142')
+        .end(done)
+    })
+
+    it('should remove other content headers', function (done) {
+      var server = http.createServer(function (req, res) {
+        var done = finalhandler(req, res)
+        res.setHeader('Content-Encoding', 'gzip')
+        res.setHeader('Content-Language', 'jp')
+        res.setHeader('Content-Range', 'bytes 0-2/10')
+        done()
+      })
+
+      request(server)
+        .get('/foo')
+        .expect(404)
+        .expect(shouldNotHaveHeader('Content-Encoding'))
+        .expect(shouldNotHaveHeader('Content-Language'))
+        .expect(shouldNotHaveHeader('Content-Range'))
+        .end(done)
+    })
+  })
+
   describe('request started', function () {
     it('should not respond', function (done) {
       var server = http.createServer(function (req, res) {