+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Remove set content headers that break response
   * deps: on-finished@2.4.1
   * deps: statuses@2.0.1
     - Rename `425 Unordered Collection` to standard `425 Too Early`